Atomic-ish organization on components
