--- src/components/Menu/MenuModal/SettingsModal/LangSelector/LangButton/styled.ts
@@ -1,23 +1,0 @@
-import styled from 'styled-components';
-import ReactCountryFlag from 'react-country-flag';
-import { TiTick } from 'react-icons/ti';
-
-import colors from '../../../../../../utils/colors';
-
-const StyledReactCountryFlag = styled(ReactCountryFlag)`
-  background-color: ${colors.activeButton};
-  padding: 10px;
-  border-radius: 50%;
-  margin-right: 1em;
-  cursor: pointer;
-`;
-
-const Tick = styled(TiTick)`
-position: absolute;
-transform: translate(-45px, 35px);
-background-color: ${colors.glassBorder};
-border-radius: 50%;
-font-size: 1.1em;
-`;
-
-export { StyledReactCountryFlag, Tick };
