+++ src/components/atoms/Counters/styled.ts
