+++ src/components/atoms/Logo/styled.ts
