+++ src/components/atoms/Timer/style.scss
