+++ src/components/atoms/Timer/styled.ts
