+++ src/components/modals/InfoModal/styled.ts
