+++ src/components/modals/SettingsModal/NumberInput/index.tsx
